Split SnapshotContent out of Data
